speed control

+++ lib/motor/motor.h
@@ -1,5 +1,7 @@
-void drive_forwards(float pwm);
-void drive_backwards(float pwm);
-void drive_right(float pwm);
-void drive_left(float pwm);
+#include <stdio.h>
+
+void drive_forwards(uint8_t pwm);
+void drive_backwards(uint8_t pwm);
+void drive_right(uint8_t pwm);
+void drive_left(uint8_t pwm);
 void motor_stop();
